Multiplayer sync code implemented, but untested
