+++ README.md
@@ -1,6 +1,6 @@
 <p align="center">
     <a href="">
-        <img src="https://img.shields.io/badge/Version-0.1.0-blue.svg?style=flat-square" alt="Zeus Enhanced Version">
+        <img src="https://img.shields.io/badge/Version-0.2.1-blue.svg?style=flat-square" alt="Zeus Enhanced Version">
     </a>
     <a href="https://steamcommunity.com/sharedfiles/filedetails/?id=1745501605">
         <img src="https://img.shields.io/badge/Steam-dev%20build-orange?logo=steam" alt="H-60 Pack">
