+++ addons/MH60S/config/cfgVehicles.hpp
@@ -13,9 +13,8 @@ class CfgVehicles {
         };
     }; // Heli_Transport_01_base_F
     class vtx_MH60S_Pylons_GAU21L: vtx_H60_base {
-        class VTX_H60_HDTS_Common;
         class VTX_H60_HDTS_Copilot;
-        class VTX_H60_HDTS_CoPilot_Turret;
+        class VTX_H60_HDTS_Pilot;
         scope = 2;
         displayName = "MH-60S Knighthawk - Pylons, GAU-21(L)";
         transportSoldier=0;
@@ -51,13 +50,18 @@ class CfgVehicles {
                 animPeriod=1;
                 initPhase=1;
             };
+            class FLIR_HIDE: MAWS_Tubes_Show {
+                initPhase=0;
+            };
+            class FLIR_BACK: FLIR_HIDE {
+                initPhase=1;
+            };
         }; // AnimationSources
         hiddenSelectionsTextures[] = {"","","","","","","","","","","","","","","","","z\vtx\addons\MH60S\data\mh60s_main_co.paa","z\vtx\addons\MH60S\data\mh60s_misc_co.paa","z\vtx\addons\MH60S\data\mh60s_tail_co.paa"};
     }; // vtx_MH60S_Pylons_GAU21L
     class vtx_MH60S_GAU21L: vtx_H60_base {
-        class VTX_H60_HDTS_Common;
         class VTX_H60_HDTS_Copilot;
-        class VTX_H60_HDTS_CoPilot_Turret;
+        class VTX_H60_HDTS_Pilot;
         scope = 2;
         displayName = "MH-60S Knighthawk - GAU-21(L)";
         transportSoldier=0;
@@ -90,13 +94,18 @@ class CfgVehicles {
                 animPeriod=1;
                 initPhase=1;
             };
+            class FLIR_HIDE: MAWS_Tubes_Show {
+                initPhase=0;
+            };
+            class FLIR_BACK: FLIR_HIDE {
+                initPhase=1;
+            };
         }; // AnimationSources
         hiddenSelectionsTextures[] = {"","","","","","","","","","","","","","","","","z\vtx\addons\MH60S\data\mh60s_main_co.paa","z\vtx\addons\MH60S\data\mh60s_misc_co.paa","z\vtx\addons\MH60S\data\mh60s_tail_co.paa"};
     }; // vtx_MH60S_GAU21L
     class vtx_MH60S_Pylons: vtx_H60_base {
-        class VTX_H60_HDTS_Common;
         class VTX_H60_HDTS_Copilot;
-        class VTX_H60_HDTS_CoPilot_Turret;
+        class VTX_H60_HDTS_Pilot;
         scope = 2;
         displayName = "MH-60S Knighthawk - Pylons";
         editorPreview = "z\vtx\addons\MH60S\Data\Preview\vtx_MH60S_Pylons.jpg";
@@ -129,13 +138,18 @@ class CfgVehicles {
                 animPeriod=1;
                 initPhase=1;
             };
+            class FLIR_HIDE: MAWS_Tubes_Show {
+                initPhase=0;
+            };
+            class FLIR_BACK: FLIR_HIDE {
+                initPhase=1;
+            };
         }; // AnimationSources
         hiddenSelectionsTextures[] = {"","","","","","","","","","","","","","","","","z\vtx\addons\MH60S\data\mh60s_main_co.paa","z\vtx\addons\MH60S\data\mh60s_misc_co.paa","z\vtx\addons\MH60S\data\mh60s_tail_co.paa"};
     }; // vtx_MH60S_Pylons_GAU21L
     class vtx_MH60S: vtx_H60_base {
-        class VTX_H60_HDTS_Common;
         class VTX_H60_HDTS_Copilot;
-        class VTX_H60_HDTS_CoPilot_Turret;
+        class VTX_H60_HDTS_Pilot;
         scope = 2;
         displayName = "MH-60S Knighthawk";
         editorPreview = "z\vtx\addons\MH60S\Data\Preview\vtx_MH60S.jpg";
