+++ addons/uh60_anvishud/config/MFD/weaponInfo.hpp
@@ -25,3 +25,31 @@ TEXT_HMD_L(WEAPON_NAME,0.015,0.81+(SMALL_LINEHEIGHT*2))
 //         }; // AMMO_NAME
 //     }; // STA2_SEL
 // };
+
+class TurretDirection {
+    type="line";
+    width = 3;
+    points[] ={
+        {"PilotCameraToView", { 0.02, -0.015}, 1},
+        {"PilotCameraToView", { 0,     0.015}, 1},
+        {"PilotCameraToView", {-0.02, -0.015}, 1},
+        {"PilotCameraToView", { 0.02, -0.015}, 1}
+    }; // points
+}; // TurretDirection
+class laserOn {
+    condition="laseron";
+    class TurretDirection {
+        type="line";
+        width = 3;
+        points[] ={
+            {"PilotCameraToView", {-0.04, -0.04}, 1},
+            {"PilotCameraToView", {-0.02, -0.02}, 1},{},
+            {"PilotCameraToView", {0.04, 0.04}, 1},
+            {"PilotCameraToView", {0.02, 0.02}, 1},{},
+            {"PilotCameraToView", {-0.04,0.04}, 1},
+            {"PilotCameraToView", {-0.02,0.02}, 1},{},
+            {"PilotCameraToView", {0.04,-0.04}, 1},
+            {"PilotCameraToView", {0.02,-0.02}, 1}
+        }; // points
+    }; // TurretDirection
+}; // laserOn
