--- addons/uh60_flir/functions/fnc_attemptSynchronization.sqf
@@ -1,1 +1,0 @@
-params ["_vehicle"];
