+++ addons/uh60_flir/functions/fnc_syncTurret.sqf
@@ -1,0 +1,7 @@
+params ["_stabilized", "_values"];
+
+if (_stabilized) then {
+	vehicle player setPilotCameraTarget _values;
+} else {
+	vehicle player setPilotCameraDirection _values;
+};
