+++ addons/uh60_hoist/$PBOPREFIX$
@@ -1,1 +1,3 @@
 z\vtx\addons\uh60_hoist
+
+version = bbdae278
